move serve_start_shutdown to conftest
